Improve debug logging
Not debug logging was done throughout the code base, and thus was not
helpful in case of problems. Now some logging has been added.

The order of main modules initialization had to be modified, esp. the
call for CommandLine class as this is the one defining the logging
level. Now is parses the logging argument independantly once init, and a
run() function has been implemented to call all the remaining commands
in case of use of the command line.