Fix location when accessed from discussion menu
When a RP is accessed through the discussion menu, the "Back" button
lead back to the menu instead of the city location - thus leading to an
error in the location parsing. By default, the location is now from the
URL if no location can be taken from the "Back" button.